Remake the player movement, remake the controls, remove the props, added slideback.
