+++ Assets/InputSystem.inputsettings.asset
@@ -12,7 +12,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: c46f07b5ed07e4e92aa78254188d3d10, type: 3}
   m_Name: InputSystem.inputsettings
   m_EditorClassIdentifier: 
-  m_SupportedDevices: []
+  m_SupportedDevices:
+  - XInputController
   m_UpdateMode: 1
   m_CompensateForScreenOrientation: 1
   m_FilterNoiseOnCurrent: 0
