--- Assets/Inputs/PlayerControls.inputactions
@@ -1,96 +1,0 @@
-{
-    "name": "PlayerControls",
-    "maps": [
-        {
-            "name": "Player",
-            "id": "380837b8-e023-4eeb-8066-b29360634218",
-            "actions": [
-                {
-                    "name": "Movement",
-                    "type": "Button",
-                    "id": "c92d5c4e-81d5-42c4-a8d8-09f6ef6dbe99",
-                    "expectedControlType": "Button",
-                    "processors": "",
-                    "interactions": "Press"
-                },
-                {
-                    "name": "SkipMovement",
-                    "type": "Button",
-                    "id": "3f47cf7f-be29-477b-a2a5-46c75b5bd277",
-                    "expectedControlType": "Button",
-                    "processors": "",
-                    "interactions": "Press"
-                }
-            ],
-            "bindings": [
-                {
-                    "name": "WS",
-                    "id": "37999aae-8441-4ec8-95fc-0a29d018bb2d",
-                    "path": "1DAxis",
-                    "interactions": "",
-                    "processors": "",
-                    "groups": "",
-                    "action": "Movement",
-                    "isComposite": true,
-                    "isPartOfComposite": false
-                },
-                {
-                    "name": "negative",
-                    "id": "6864389b-f1e0-4b4a-84ff-8a0d8f417a7c",
-                    "path": "<Keyboard>/w",
-                    "interactions": "",
-                    "processors": "",
-                    "groups": "",
-                    "action": "Movement",
-                    "isComposite": false,
-                    "isPartOfComposite": true
-                },
-                {
-                    "name": "positive",
-                    "id": "f9ea55d8-8910-4167-80a3-0251c41caf06",
-                    "path": "<Keyboard>/s",
-                    "interactions": "",
-                    "processors": "",
-                    "groups": "",
-                    "action": "Movement",
-                    "isComposite": false,
-                    "isPartOfComposite": true
-                },
-                {
-                    "name": "1D Axis",
-                    "id": "c83e0782-0f4c-47ba-a62d-2cf35d703ca2",
-                    "path": "1DAxis",
-                    "interactions": "",
-                    "processors": "",
-                    "groups": "",
-                    "action": "SkipMovement",
-                    "isComposite": true,
-                    "isPartOfComposite": false
-                },
-                {
-                    "name": "negative",
-                    "id": "34c989bb-1470-4d7a-9630-1f9215ff2705",
-                    "path": "<Keyboard>/a",
-                    "interactions": "",
-                    "processors": "",
-                    "groups": "",
-                    "action": "SkipMovement",
-                    "isComposite": false,
-                    "isPartOfComposite": true
-                },
-                {
-                    "name": "positive",
-                    "id": "f927a597-72b4-4b6d-aee7-d4246ae92fc0",
-                    "path": "<Keyboard>/d",
-                    "interactions": "",
-                    "processors": "",
-                    "groups": "",
-                    "action": "SkipMovement",
-                    "isComposite": false,
-                    "isPartOfComposite": true
-                }
-            ]
-        }
-    ],
-    "controlSchemes": []
-}
