--- Assets/Inputs/PlayerControls.inputactions.meta
@@ -1,14 +1,0 @@
-fileFormatVersion: 2
-guid: 88b40957be9449c4bb278c775ca69b55
-ScriptedImporter:
-  internalIDToNameTable: []
-  externalObjects: {}
-  serializedVersion: 2
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-  script: {fileID: 11500000, guid: 8404be70184654265930450def6a9037, type: 3}
-  generateWrapperCode: 1
-  wrapperCodePath: 
-  wrapperClassName: 
-  wrapperCodeNamespace:
