--- Assets/Movement/PlayerMovementBehavior.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8d7dd5edcfeea8d4088f5bb560a4d73a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
