+++ Assets/Scripts/Gameplay/Serving.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1d88ea92b18383d4c9d65657ed8abe80
+guid: 455435c0aa5b2ba45b0eb00298b67c1e
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
