+++ Assets/Scripts/Gameplay/Serving/SlideBackBehavior.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5740aae8db4e3164699e5f18f0cc9d5e
+guid: 6f9b92d5ea591f042a025048a72c73c4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
